SAK-7431 eliminated java 1.5 syntax
git-svn-id: https://source.sakaiproject.org/svn/content/branches/sakai_2-3-x@19878 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ content-tool/tool/src/java/org/sakaiproject/content/tool/ResourcesAction.java
@@ -10550,8 +10550,8 @@ public class ResourcesAction
 			{
 				expansionLimit = new Integer(EXPANDABLE_FOLDER_SIZE_LIMIT);
 			}
-			folder.setSortable(ContentHostingService.isSortByPriorityEnabled() && collection_size > 1 && collection_size <= expansionLimit);
-			folder.setIsTooBig(collection_size > expansionLimit);
+			folder.setSortable(ContentHostingService.isSortByPriorityEnabled() && collection_size > 1 && collection_size <= expansionLimit.intValue());
+			folder.setIsTooBig(collection_size > expansionLimit.intValue());
 				
 			folder.setDepth(depth);
 			newItems.add(folder);
